Slightly better.. I thikn
