+++ handlers/bookmarks/routes/bookmarkspage.go
@@ -1,0 +1,10 @@
+package routes
+
+import (
+	"net/http"
+)
+
+func BookmarksPage(w http.ResponseWriter, r *http.Request) {
+	t := NewBookmarksTask().(*bookmarksTask)
+	t.Get(w, r)
+}
