+++ handlers/bookmarks/routes/mine.go
@@ -1,12 +1,14 @@
-package bookmarks
+package routes
 
 import (
 	"database/sql"
 	"errors"
-	"github.com/arran4/goa4web/internal/tasks"
 	"log"
 	"net/http"
 	"strings"
+
+	"github.com/arran4/goa4web/handlers/bookmarks/internal"
+	"github.com/arran4/goa4web/internal/tasks"
 
 	"github.com/arran4/goa4web/core/common"
 	"github.com/arran4/goa4web/core/consts"
@@ -16,19 +18,19 @@ func MinePage(w http.ResponseWriter, r *http.Request) {
 	cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
 	cd.PageTitle = "My Bookmarks"
 
-	var cols []*Column
+	var cols []*internal.Column
 	bm, err := cd.Bookmarks()
 	if err != nil && !errors.Is(err, sql.ErrNoRows) {
 		log.Printf("error getBookmarksForUser: %s", err)
 	} else if bm != nil {
 		list := strings.TrimSpace(bm.List.String)
 		if list != "" {
-			cols = ParseColumns(list)
+			cols = internal.ParseColumns(list)
 		}
 	}
 
 	MinePageTmpl.Handle(w, r, struct {
-		Columns []*Column
+		Columns []*internal.Column
 	}{cols})
 }
