+++ handlers/bookmarks/routes/mine_test.go
@@ -1,4 +1,4 @@
-package bookmarks
+package routes
 
 import (
 	"database/sql"
@@ -7,7 +7,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/google/go-cmp/cmp"
 	"github.com/gorilla/sessions"
 
 	"github.com/arran4/goa4web/core"
@@ -16,56 +15,6 @@ import (
 	"github.com/arran4/goa4web/internal/db"
 	"github.com/arran4/goa4web/internal/testhelpers"
 )
-
-func TestParseColumns(t *testing.T) {
-	tests := []struct {
-		name      string
-		bookmarks string
-		want      []*Column
-	}{
-		{
-			name:      "Test",
-			bookmarks: "Category: Search\nhttp://www.google.com.au Google\nCategory: Wikies\nhttp://en.wikipedia.org/wiki/Main_Page Wikipedia\nhttp://mathworld.wolfram.com/ Math World\nhttp://gentoo-wiki.com/Main_Page Gentoo-wiki\n",
-			want: []*Column{{
-				Categories: []*Category{
-					{
-						Name: "Search",
-						Entries: []*Entry{
-							{
-								Url:  "http://www.google.com.au",
-								Name: "Google",
-							},
-						},
-					},
-					{
-						Name: "Wikies",
-						Entries: []*Entry{
-							{
-								Url:  "http://en.wikipedia.org/wiki/Main_Page",
-								Name: "Wikipedia",
-							},
-							{
-								Url:  "http://mathworld.wolfram.com/",
-								Name: "Math World",
-							},
-							{
-								Url:  "http://gentoo-wiki.com/Main_Page",
-								Name: "Gentoo-wiki",
-							},
-						},
-					},
-				}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := ParseColumns(tt.bookmarks)
-			if diff := cmp.Diff(got, tt.want); diff != "" {
-				t.Errorf("ParseColumns() = diff\n%s", diff)
-			}
-		})
-	}
-}
 
 func TestMinePage(t *testing.T) {
 	t.Run("Happy Path", func(t *testing.T) {
