+++ handlers/bookmarks/routes/tasks.go
@@ -1,4 +1,4 @@
-package bookmarks
+package routes
 
 import "github.com/arran4/goa4web/internal/tasks"
