--- handlers/bookmarks/section.go
@@ -1,6 +1,0 @@
-package bookmarks
-
-const (
-	// SectionWeight controls the order of the Bookmarks section in navigation.
-	SectionWeight = 70
-)
