+++ handlers/bookmarks/tasks_register.go
@@ -1,11 +1,14 @@
 package bookmarks
 
-import "github.com/arran4/goa4web/internal/tasks"
+import (
+	"github.com/arran4/goa4web/handlers/bookmarks/routes"
+	"github.com/arran4/goa4web/internal/tasks"
+)
 
 // RegisterTasks returns bookmark related tasks.
 func RegisterTasks() []tasks.NamedTask {
 	return []tasks.NamedTask{
-		saveTask,
-		createTask,
+		routes.SaveTask,
+		routes.CreateTask,
 	}
 }
